Plugin to process LLC switch info from system name
Some Neutron drivers expect the switch_info field in a port's
local_link_connection attribute to contain the system name of a
switch.  This plugin will store the system name received via LLDP (if
present) in the switch_info field of the Ironic port's
local_link_connection attribute.

It should be noted that some Neutron mechanism drivers expect
switch_info to contain something other than the system name, in which
case this plugin should not be used.

Change-Id: I95f8cad6f4f9549369d43cba1864047d461f50ad